Big update

+++ src/main/java/de/kniffo80/mobplugin/entities/monster/jumping/MagmaCube.java
@@ -43,7 +43,7 @@ public class MagmaCube extends WalkingMonster {
     }
 
     @Override
-    public int getMaxJumpHeight() {
+    public float getMaxJumpHeight() {
         return 2;
     }
