+++ src/main/java/de/kniffo80/mobplugin/entities/monster/jumping/Slime.java
@@ -43,7 +43,7 @@ public class Slime extends WalkingMonster {
     }
 
     @Override
-    public int getMaxJumpHeight() {
+    public float getMaxJumpHeight() {
         return 2;
     }
